+++ package.json
@@ -10,14 +10,17 @@
   "scripts": {
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:e2e": "jest -c jest.e2e.config.js",
-    "test:e2e:dev": "env E2E_TARGET=dev jest -c jest.e2e.config.js",
-    "test:e2e:qa": "env E2E_TARGET=qa jest -c jest.e2e.config.js",
+    "test:e2e": "jest -c test/jest.e2e.config.js",
+    "test:e2e:dev": "env E2E_TARGET=dev jest -c test/jest.e2e.config.js",
+    "test:e2e:qa": "env E2E_TARGET=qa jest -c test/jest.e2e.config.js",
     "lint": "eslint . lib",
-    "deploy:qa": "bash ./bin/deploy-qa.sh",
-    "shutdown:qa": "bash ./bin/remove-qa.sh",
-    "deploy:dev": "bash ./bin/deploy-dev.sh",
+    
     "deploy": "bash ./bin/deploy-dev.sh",
+    "deploy:qa": "bash ./bin/deploy-qa.sh",
+    "deploy:dev": "bash ./bin/deploy-dev.sh",
+    
+    "shutdown:qa": "bash ./bin/remove-qa.sh",
+    "shutdown:dev": "bash ./bin/remove-dev.sh",
     "logs": "serverless logs -f getAbapVersionShieldJson --stage dev -t"
   },
   "author": "Alexander Tsybulsky",
