+++ test/lambda.e2e.test.js
@@ -32,6 +32,7 @@ test('should process abap constant', async () => {
 });
 
 test.skip('should process apack', async () => {
+    // 2021-08
     // skip because abap-platform-jak sample does not use proper semver (X.Y is not valid, only X.Y.Z)
     // and I didn't find proper stable example in the web, seems APACK does not took off really
     const resp = await fetch(getUrl('github/SAP-samples/abap-platform-jak/.apack-manifest.xml'));
