+++ test/localtest-const.js
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 
-const handler = require('./handler');
+const handler = require('../handler');
 
 (async function main() {
     const event = {
         resource: '/version-shield-json/{sourcePath}',
-        path: '/version-shield-json/github/SAP-samples/abap-platform-jak/.apack-manifest.xml',
+        path: '/version-shield-json/github/sbcgua/mockup_loader/src/zif_mockup_loader.intf.abap/version',
         pathParameters: {
-            sourcePath: 'github/SAP-samples/abap-platform-jak/.apack-manifest.xml'
+            sourcePath: 'github/sbcgua/mockup_loader/src/zif_mockup_loader.intf.abap/version'
         },
     };
     const context = {
